+++ DSCResources/MSFT_SmbShare/MSFT_SmbShare.psm1
@@ -56,6 +56,7 @@ function Get-TargetResource
         ShareType             = [System.String] $null
         ShadowCopy            = $false
         Special               = $false
+        ScopeName             = [System.String] $null
     }
 
     $accountsFullAccess   = [system.string[]] @()
@@ -80,6 +81,7 @@ function Get-TargetResource
         $returnValue['ShareType'] = $smbShare.ShareType.ToString()
         $returnValue['ShadowCopy'] = $smbShare.ShadowCopy
         $returnValue['Special'] = $smbShare.Special
+        $returnValue['ScopeName'] = $smbShare.ScopeName
 
         $currentSmbShareAccessPermissions = Get-SmbShareAccess -Name $Name
 
@@ -183,6 +185,13 @@ function Get-TargetResource
 
     .PARAMETER Ensure
         Specifies if the SMB share should be added or removed.
+
+    .PARAMETER ScopeName
+        Specifies the scope in which the share should be created.
+
+    .PARAMETER Force
+        Specifies if the SMB share is allowed to be dropped and recreated (required
+        when the path changes).
 #>
 function Set-TargetResource
 {
@@ -242,7 +251,15 @@ function Set-TargetResource
         [Parameter()]
         [ValidateSet('Present', 'Absent')]
         [System.String]
-        $Ensure = 'Present'
+        $Ensure = 'Present',
+
+        [Parameter()]
+        [System.String]
+        $ScopeName = '*',
+
+        [Parameter()]
+        [System.Boolean]
+        $Force
     )
 
     Assert-AccessPermissionParameters @PSBoundParameters
@@ -261,11 +278,36 @@ function Set-TargetResource
 
         if ($Ensure -eq 'Present')
         {
+            if (
+                ($currentSmbShareConfiguration.Path -ne $Path -or
+                $currentSmbShareConfiguration.ScopeName -ne $ScopeName) -and
+                $Force
+            )
+            {
+                Write-Verbose -Message ($script:localizedData.RecreateShare -f $Name)
+
+                try
+                {
+                    Remove-SmbShare -Name $Name -Force -ErrorAction Stop
+                    New-SmbShare -Name $Name -Path $Path -ErrorAction Stop
+                }
+                catch
+                {
+                    Write-Error -Message ($script:localizedData.RecreateShareError -f $Name, $_)
+                }
+            }
+            else
+            {
+                Write-Warning -Message (
+                    $script:localizedData.NoRecreateShare -f $Name, $currentSmbShareConfiguration.Path, $Path
+                )
+            }
+
             Write-Verbose -Message $script:localizedData.UpdatingProperties
 
             $parametersToRemove = $smbShareParameters.Keys |
                 Where-Object -FilterScript {
-                    $_ -in ('ChangeAccess','ReadAccess','FullAccess','NoAccess','Ensure','Path')
+                    $_ -in ('ChangeAccess','ReadAccess','FullAccess','NoAccess','Ensure','Path','Force')
                 }
 
             $parametersToRemove | ForEach-Object -Process {
@@ -316,6 +358,7 @@ function Set-TargetResource
         if ($Ensure -eq 'Present')
         {
             $smbShareParameters.Remove('Ensure')
+            $smbShareParameters.Remove('Force')
 
             Write-Verbose -Message ($script:localizedData.CreateShare -f $Name)
 
@@ -385,6 +428,13 @@ function Set-TargetResource
 
     .PARAMETER Ensure
         Specifies if the SMB share should be added or removed.
+
+    .PARAMETER ScopeName
+        Specifies the scope in which the share should be created.
+
+    .PARAMETER Force
+        Specifies if the SMB share is allowed to be dropped and recreated (required
+        when the path changes).
 #>
 function Test-TargetResource
 {
@@ -445,8 +495,18 @@ function Test-TargetResource
         [Parameter()]
         [ValidateSet('Present', 'Absent')]
         [System.String]
-        $Ensure = 'Present'
+        $Ensure = 'Present',
+
+        [Parameter()]
+        [System.String]
+        $ScopeName = '*',
+
+        [Parameter()]
+        [System.Boolean]
+        $Force
     )
+
+    $null = $PSBoundParameters.Remove('Force')
 
     Assert-AccessPermissionParameters @PSBoundParameters